order/app/requirements.txt
fastapi[standard]
pika
mysql-connector-python
